// Carpooling/src/app/model/enums.model.ts
export enum UserRole {
  Passenger = 0,
  Driver = 1,
  Admin = 2
}

export enum Gender {
  Any = 2,
  Male = 0,
  Female = 1
}

export enum TripStatus {
  Pending = 'Pending',
  Confirmed = 'Confirmed',
  Ongoing = 'Ongoing',
  Completed = 'Completed',
  Cancelled = 'Cancelled'
}

export enum JoinStatus {
  Pending = 0,
  Approved = 1,
  Rejected = 2
}

export enum VerificationStatus {
  Pending = 'Pending',
  Approved = 'Approved',
  Rejected = 'Rejected'
} 